refactor(home): extract provider topology derivation
